--- tg-portfolio/components/Portfolio.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Portfolio = () => (
-  <div>Portfolio</div>
-);
-
-export default Portfolio;
