+++ tg-portfolio/components/index.js
@@ -1,14 +1,13 @@
-export { default as Navbar } from './Navbar';
+export { default as Navbar } from './layyout/Navbar';
 export { default as Sidebar } from './Sidebar';
-export { default as Footer } from './Footer';
-export { default as About } from './About';
-export { default as Services } from './Services';
-export { default as Programming } from './Programming';
-export { default as Education } from './Education';
-export { default as Testimonials } from './Testimonials';
+export { default as Footer } from './layyout/Footer';
+export { default as About } from './about/About';
+export { default as Services } from './about/Services';
+export { default as Programming } from './skills/Programming';
+export { default as Education } from './skills/Education';
+export { default as Testimonials } from './index/Testimonials';
 export { default as Contact } from './Contact';
 export { default as Projects } from './Projects';
-export { default as Hero } from './Hero';
-export { default as Portfolio } from './Portfolio';
-export { default as Experience } from './Experience';
+export { default as Hero } from './index/Hero';
+export { default as Experience } from './skills/Experience';
