+++ tg-portfolio/pages/_app.js
@@ -1,6 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import '../styles/globals.css';
-import Layout from '../components/Layout';
+import Layout from '../components/layyout/Layout';
 
 const MyApp = ({ Component, pageProps }) => (
   <Layout>
